Add option to validate settings values
When defining a LocalSetting, a validator function can now be passed.
Such a function takes a value and returns True or False. Currently,
this is only used to validate missing values that are prompted for.
It's assumed that if a value is read from a config file that it's valid
(i.e., it was either written there by the Checker or the user typed in
the value they want). It's also assumed that defaults are valid even if
they wouldn't pass the validator.

XXX: Not sure if values should also be validated when they're loaded...