Fix relative path issue when extending
When a settings file (the extender) specifies that it extends another
file (the extended) and the path to the extended file is relative, the
extended file's path should be made absolute relative to the directory
containing the extender. Otherwise, things will break when $PWD isn't
the directory containing the files.