Fix/improve loading of lists
This for settings specified like `TEMPLATES.0.XYZ`.

The previously implementation was horrendously complex and also buggy.

This new version is much simpler and actually works.

I had considered removing this functionality altogether, but it's
actually quite useful for things like Django 1.8's new `TEMPLATES`
settings (which is a list of template configurations).