Update basic usage section of README
In particular, the way local settings are loaded was changed a while
back (in de95645) and that was not reflected in the README.

In addition, add info about imports and a bit more info and
clarification of how local settings are loaded.

Credit: kfarr2 made a PR to document in the README which imports are
needed. That's when I noticed the other outdated info.