Don't explode when wrong settings module is guessed
When running the make-local-settings script, if no base settings module
is specified, we try to guess, which won't work if the distribution name
is different from the main top-level package name.

Introduce a better method for guessing the base settings module that
doesn't rely on the distribution name being the same as the main
top-level package name.

This method also avoids doing any imports to try to guess the base
settings module, because that could have unintended side effects.