In make_local_settings, load settings as an app would
That is, simply import the base settings module, which we assume does a
call to load_and_check_settings (and if it doesn't... well, that won't
work--maybe add a flag for this).

This makes the script work when a base settings module has settings
that are derived from or depend on local settings being loaded. E.g.,
before, make_local_settings would explode in this situation:

    PANTS = LocalSetting('pants')
    ...
    globals().update(load_and_check_settings(globals()))
    ...
    if PANTS == 'jeans':
        ...