Import indicated settings
The values of the settings listed in IMPORT_FROM_STRING will be imported
using Django's import_string utility (if a value isn't a string, it will
be used as is).

This makes it easy to specify some dynamic settings via local settings
without having to jump back into the settings module.