Append EXTRA settings
Any setting specified with the `EXTRA` prefix will be appended to the
corresponding setting. E.g., if `EXTRA.INSTALLED_APPS` is specified, it
will be appended to `INSTALLED_APPS`.

This is another attempt at making it easy to specify all settings via
local settings.

TODO: This could use a bit of type checking in order to provide user
      friendly error messages.