Rewrite Loader.load()
Most importantly, this fixes a bug found when adding tests in 00e95e2.

In addition, we now use `OrderedDict`s internally, not just for the top
level settings object. This is another attempt at making interpolation
more deterministic.

The new implementation is also quite a bit cleaner, especially since it
removes a hacky workaround for the aforementioned bug.