Make Loader._import_from_string more robust
Return early if IMPORT_FROM_STRING is not set, empty, or set to None.
Previously, when IMPORT_FROM_STRING was explicitly set to None, that
would cause an error when trying to iterate over None. We want to allow
IMPORT_FROM_STRING to be set to None either as a default value or as a
way to disable imports specified in base settings.