Make Loader._append_extras more robust
Return early if EXTRA is not set, empty, or set to None. Previously,
when EXTRA was explicitly set to None, that would cause an error when
trying to iterate over None. We want to allow EXTRA to be set to None
either as a default value or as a way to disable imports specified in
base settings.

Allow the value of an EXTRA to be None for similar reasons. This would
be something like setting `EXTRA.INSTALLED_APPS = None`.

Throw a TypeError if a setting specified in EXTRA isn't a sequence type.