Add SWAP post processor
This provides a generic mechanism for swapping out items in list-type
settings. E.g., if `INSTALLED_APPS = ["a", "b", "c"]`, it's now possible
to swap app "a" by doing `SWAP.INSTALLED_APPS = {"a": "alternative"}`.

It's already possible to do `INSTALLED_APPS.0 = "alternative"`, but that
can be brittle since the order of `INSTALLED_APPS` in the base settings
could change.

The other option is to redefine `INSTALLED_APPS`, but that can be
tedious when you only need to change one thing.

Note that we don't need this mechanism for dict-type settings because
doing `XYZ.key` works well and doesn't have the brittleness problem.