When interpolating, ignore missing keys
This amends 2c06cdc. It's necessary for cases where a setting value
contains "{xyz}" where "{xyz}" isn't meant to be interpolated. Before,
we could get around this by doing "{{xyz}}", but that doesn't work any
more because of multiple rounds of interpolation.

The real fix here would be to use a different syntax for settings that
should be interpolate vs normal Python format strings.