Use strategies for reading & writing settings files
This pulls the code for reading config files out of `Loader` and the
code for writing config files out of `Checker`. This functionlity is now
implemented via strategy classes. Currently, there is only one strategy:
the `INIJSONStrategy`. Other strategies should be fairly straightforward
to implement now (e.g., a `YAMLStrategy`).

This also pulls the logic for traversing settings into a generic
`Settings` class that should be usable regardless of what type of file
settings are read from. This `Settings` class is slightly enhanced over
a built-in dict in that it allows access to items via both item and
attribute access. It's also possible to use dotted names to fetch nested
settings.

One notable effect of these changes is that `Loader` is now simpler
because it uses a higher level interface internally to manipulate
settings instead of using traversal directly, which was somewhat
confusing.

Some additional tests were added to aid in refactoring.