Make dotted item access explicit
That is, instead of automatically treating dotted names as references to
nested items when using normal item access, a new set of methods
explicitly provides that behavior:

	- Settings.contains_dotted()
	- Settings.get_dotted()
	- Settings.set_dotted()

This makes the use of settings outside of the django-local-settings
package itself more normal and removes some gotchas and surprising
behavior. For example, to add an item with the key '()' before required
adding the key '(())'. And when iterating, keys with dots in them would
be wrapped in parentheses, which isn't always desirable (i.e., when
using such keys outside of this package).

This also makes the implementation of Settings *much* simpler since we
don't have to worry about recursive traversal now or any weirdness with
parentheses.