Add/declare support for Django 1.10
This didn't require any code changes, just a mention in the README and
adding Django 1.10 to the tox config.

Also, fix up the env list in tox.ini so that all Django versions for a
given Python version will be tested in order. I think this will also be
easier to maintain.