Guess strategy type from file name when not specified
Previously, the `INIJSONStrategy` would *always* be used if a strategy
type wasn't specified. Now we discover the strategy type based on the
local settings file's type/extension.

In order to support this, the bulk of the logic in the
`Strategy.parse_file_name_and_section` method was moved into a new
utility function in the `util` module.