Simplify internals of interpolation in Loader
This replaces `Loader._interpolate_values_inner()` with
`Loader._traverse_object()`, which simplify traverses an object and
applies an action to each string encountered. This will allow it to be
used for purposes other than interpolation.

`Loader._interpolate_values()` now manages its own interpolation list
instead of sending that down to the traversal method, which is much
cleaner.