When reading, mark values that can't be decoded
This trickles down to big simplifications in `Loader`.

First, it no longer attempts to decode values read from the settings
file; this is handled internally in the strategy where it belongs.

Because the strategy marks values that couldn't be decoded, we no longer
need to do that in `Loader`.

And instead of doing two rounds of interpolation, we now do one round of
interpolation and then do a second pass through the settings to convert
all the values that previously couldn't be decoded (decoding errors at
this stage will be propagated).