Fix/improve settings file inheritance
Consider the following settings file:

    # local.base.cfg
    [DEFAULT]
    DEBUG = false

    [dev]
    DEBUG = true
    X = 1

    [other]
    extends = "#dev"

Previously, when loading settings from the `other` section, the value of
`DEBUG` would be `False` instead of the expected `True`. That's because
of the way the standard library's `configparser.RawConfigParser` reads
items from a section: items from the default section are included by
default.

So, when reading the `other` section, its items would effectively be
`{'DEBUG': False}`, which would override `{'DEBUG': False}` from the
`dev` section.

Now, when reading the `other` section, the `DEBUG` value from the `dev`
section will take precedence over the default `DEBUG` value. This is
accomplished by loading the defaults separately and reading section
items directly without defaults and then merging the defaults and all
the sections in order.

One other tricky aspect of this is how to handle defaults in one file
when it extends from another file. The question here is, should the
*defaults* from the first file (the one that's doing the extending) take
precedence over the items from the *section* being read in the other
file? I've decided they should.

The following illustrates this:

    # local.cfg
    [DEFAULT]
    extends = "local.base.cfg"
    X = "higher precedence"

    [dev]
    SOME_OTHER_SETTING = 1

When reading the `dev` section from `local.cfg`, the value of `X` will
be `"higher precedence"` even though `X` is also defined in the `dev`
section of `local.base.cfg`.

- Add abstract method `Strategy.get_defaults()`.
- Change interface of abstract method `Strategy.read_section()` so that
  it returns either the section's items without defaults if the section
  is present or an empty dict otherwise.
- Modify interface of `Strategy.read_file()` so that it returns only the
  settings instead of both the settings and the section present flag.
- Add `Strategy._read_one_file()`.
- Add `LocalSettingsConfigParser.get_section()` for getting a section's
  items without defaults.
- Add `INIStrategy.get_defaults()`.
- Update `INIStrategy.read_section()`.