When creating virtualenv, upgrade setuptools
setuptools isn't listed as an explicit dependency in the pip package, so
it's not updated when using `pip install -U --upgrade-strategy eager`
like I thought it would be.

pip apparently uses whatever setuptools version happens to be installed,
which when using `python -m venv` to create a virtualenv will likely be
out of date.

This can cause problems when creating releases, for example, because the
package metadata might not match what PyPI is expecting.