Fix bug in Checker.check() wrt. tracking seen settings
This method keeps of settings that were already seen and processed so
settings aren't processed twice.

The previous version kept track of these by partial key (with prefix
stripped) instead of the entire key.

So, if a top level setting and a subsetting had the same name, one of
them would be inadvertently skipped (depending on dict iteration order):

    XYZ = 1
    ABC.XYZ = 1