Add initial support for env settings
This allows local settings to be specified in the environ and is an
alternative to `SecretSetting`. Env vars will be loaded from a .env
file, if present (`./.env` by default).

This functionality can be used without a local settings file by passing
`env_only=True` to `load_and_check_settings()`.