Add support for dates, times, and additional number syntaxes
Examples:

- 0b1111
- 0o76543210
- 0xfedcba9876543210
- +1
- 1_000_000
- 1_000_000.0
- 2021-06-23
- 2021-06-23T12:30
- 2021-06-23T12:30:00.000000
- 12:30

Basically, support the same values as TOML. One deficiency here for the
time being is that this special parsing only works for non-nested
values. The `json` package in the standard library seemingly makes it
impossible to access raw string values that can't be parsed (i.e., a
hook that would be analogous to the `default` arg to `json.dumps()`).

A related change is that when prompting the user for missing local
settings, the values they type in will be *validated* using the JSON
encoder but will be inserted into the settings file exactly as they
typed. This keeps typed values like `0xbb00` from being converted to
`47872`.