[jsonish] Wrap internal scan function in `make_scan_method`
This localizes all the `Scanner`'s instance config in the scan function.
In particular, we want to avoid attribute and global lookups because
these can really add up. A general rule is that there should be no dot
access except to locals, and those should be avoided as well by
assigning heavily used attributes to other locals.

A related TODO item will be avoiding function calls as much as possible.
Currently, the whitespace consuming function is called frequently, and
this should probably be optimized at some point.